+++ editor/src/Application.cpp
@@ -84,7 +84,7 @@ namespace editor {
             RB_EDITOR_INFO("Window's new size = {0} and {1}", evt.size.widht, evt.size.heigth);
             m_window -> resize({static_cast<int>(evt.size.widht),
                                 static_cast<int>(evt.size.heigth)});
-            m_window -> setView({{0, 0}, {evt.size.widht, evt.size.heigth}});
+            m_window -> setView({{0, 0}, {static_cast<float>(evt.size.widht), static_cast<float>(evt.size.heigth)}});
         });
 
         m_eventBinder.bindEvent(robot2D::Event::KeyPressed, [this](const robot2D::Event& evt) {});
