+++ editor/src/Editor.cpp
@@ -456,7 +456,7 @@ namespace editor {
         if(event.mouse.btn == robot2D::mouse2int(robot2D::Mouse::MouseLeft)
                         && !m_interactor -> hasSelectedEntities() && !m_guizmo2D.isActive() && viewportPanel.isActive()) {
             m_leftMousePressed = true;
-            auto mousePos = m_editorCamera -> convertPixelToCoords({event.mouse.x, event.mouse.y});
+            auto mousePos = m_editorCamera -> convertPixelToCoords({static_cast<float>(event.mouse.x), static_cast<float>(event.mouse.y)});
             m_selectionCollider.setPosition(mousePos);
             m_selectionCollider.setIsShown(true);
         }
